requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Configuration
pyyaml>=6.0.1
watchdog>=3.0.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
